:truck: Move trav-com composables to the appropriate location.
